add coverage for all custom elements
